requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-multipart==0.0.20

# Database
aiosqlite==0.20.0

# Data validation
pydantic==2.10.4

# Browser Automation
playwright==1.49.1

# LLM Providers
openai==1.58.1
anthropic==0.42.0

# Configuration
python-dotenv==1.0.1

# Logging
rich==13.9.4
